+++ trunk/windows/PopupManager.cpp
@@ -102,7 +102,7 @@ void PopupManager::Show(const tstring &aMsg, const tstring &aTitle, int Icon, in
 	popups.push_back(p);
 }
 
-void PopupManager::on(TimerManagerListener::Second /*type*/, uint32_t /*tick*/) {
+void PopupManager::on(TimerManagerListener::Second /*type*/, uint64_t /*tick*/) {
 	//post a message and let the main window thread take care of the window
 	::PostMessage(WinUtil::mainWnd, WM_SPEAKER, MainFrame::REMOVE_POPUP, 0);
 }
